logger: move `devFormatter` to a special file
